+++ YARG.Core/Chart/Sync/SyncTrack_FW.cs
@@ -23,12 +23,12 @@ namespace YARG.Core.Chart
         }
 
         internal const int MICROS_PER_SECOND = 1000000;
-        public float ConvertToSeconds(long ticks, int startIndex = 0)
+        public double ConvertToSeconds(long ticks, int startIndex = 0)
         {
             return ConvertToSeconds(ticks, ref startIndex);
         }
 
-        public float ConvertToSeconds(long ticks, ref int startIndex)
+        public double ConvertToSeconds(long ticks, ref int startIndex)
         {
             var span = _tempoMarkers!.Span;
             int length = span.Length;
@@ -38,22 +38,22 @@ namespace YARG.Core.Chart
                 {
                     ref var marker = ref span[i];
                     startIndex = i;
-                    return ((marker.obj.Micros * (ticks - marker.position) / (float) _tickrate) + marker.obj.Anchor) / MICROS_PER_SECOND;
+                    return ((marker.obj.Micros * (ticks - marker.position) / (double) _tickrate) + marker.obj.Anchor) / MICROS_PER_SECOND;
                 }
             }
             throw new Exception("dafuq");
         }
 
-        public long ConvertToTicks(float seconds, int startIndex = 0)
+        public long ConvertToTicks(double seconds, int startIndex = 0)
         {
             return ConvertToTicks(seconds, ref startIndex);
         }
 
-        public long ConvertToTicks(float seconds, ref int startIndex)
+        public long ConvertToTicks(double seconds, ref int startIndex)
         {
             var span = _tempoMarkers!.Span;
             int length = span.Length;
-            float micros = seconds * MICROS_PER_SECOND;
+            double micros = seconds * MICROS_PER_SECOND;
             for (int i = startIndex; i < length; i++)
             {
                 if (i + 1 == length || micros < span[i + 1].obj.Anchor)
